Silver Tier/requirements.txt
watchdog==4.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dateutil==2.8.2
selenium==4.15.2
webdriver-manager==4.0.1
